+++ doc/release-notes.md
@@ -5,3 +5,5 @@ Bitcoin ABC version 0.24.2 is now available from:
   <https://download.bitcoinabc.org/0.24.2/>
 
 This release includes the following features and fixes:
+ - The message signature prefix has been rebranded to `eCash Signed Message:`
+   (previously was `Bitcoin Signed Message:`).